move checksum test
