+++ chemreg/compound/tests/test_utils.py
@@ -1,5 +1,4 @@
 import re
-from random import randint
 
 from django.core.cache import cache
 
@@ -7,23 +6,7 @@ import pytest
 
 from chemreg.compound.models import BaseCompound
 from chemreg.compound.settings import compound_settings
-from chemreg.compound.utils import build_cid, compute_checksum, extract_int
-
-
-def test_compute_checksum():
-    i = randint(2000000, 9999999)
-    computed = (
-        (1 * int(str(i)[0]))
-        + (2 * int(str(i)[1]))
-        + (3 * int(str(i)[2]))
-        + (4 * int(str(i)[3]))
-        + (5 * int(str(i)[4]))
-        + (6 * int(str(i)[5]))
-        + (7 * int(str(i)[6]))
-    ) % 10
-    checksum = compute_checksum(i)
-    assert computed == checksum
-    assert 0 <= checksum < 10
+from chemreg.compound.utils import build_cid, extract_int
 
 
 @pytest.mark.django_db
